+++ module/bukkit-nms/bukkit-nms-data-component/src/main/kotlin/taboolib/module/nms/component/internal/EnchantmentGlintOverrideType.kt
@@ -1,4 +1,4 @@
-package taboolib.module.nms.component.types
+package taboolib.module.nms.component.internal
 
 import net.minecraft.core.component.DataComponents
 import net.minecraft.world.item.ItemStack
@@ -7,36 +7,33 @@ import taboolib.module.nms.MinecraftVersion
 import taboolib.module.nms.component.ComposedType
 
 /**
- * HideTooltipType - 完全隐藏物品提示框（1.20.5+）
+ * EnchantmentGlintOverrideType - 附魔光效强制覆盖（1.20.5+）
  *
- * 存在该组件时，鼠标悬停不会显示任何提示信息
+ * true  = 强制显示附魔光效（无附魔时也显示）
+ * false = 强制隐藏附魔光效（有附魔时也不显示）
+ * null  = 使用默认行为（有附魔则显示）
  *
  * @author TheFloodDragon
  * @since 2026/2/19
  */
 @Suppress("unused")
-class HideTooltipType : ComposedType<Boolean>() {
+class EnchantmentGlintOverrideType : ComposedType<Boolean>() {
 
     override fun get(item: Any): Boolean? {
         if (MinecraftVersion.versionId >= 12005) {
-            return if ((item as ItemStack).has(DataComponents.HIDE_TOOLTIP)) true else null
+            return (item as ItemStack).get(DataComponents.ENCHANTMENT_GLINT_OVERRIDE)
         } else throw UnsupportedVersionException()
     }
 
     override fun set(item: Any, value: Boolean) {
-        item as ItemStack
         if (MinecraftVersion.versionId >= 12005) {
-            if (value) {
-                item.set(DataComponents.HIDE_TOOLTIP, net.minecraft.util.Unit.INSTANCE)
-            } else {
-                item.remove(DataComponents.HIDE_TOOLTIP)
-            }
+            (item as ItemStack).set(DataComponents.ENCHANTMENT_GLINT_OVERRIDE, value)
         } else throw UnsupportedVersionException()
     }
 
     override fun remove(item: Any) {
         if (MinecraftVersion.versionId >= 12005) {
-            (item as ItemStack).remove(DataComponents.HIDE_TOOLTIP)
+            (item as ItemStack).remove(DataComponents.ENCHANTMENT_GLINT_OVERRIDE)
         } else throw UnsupportedVersionException()
     }
